Refactor: update FrameworkInfo to AboutView to align with Model and Page, rename and reorganize files in 'pages' folder
