+++ QualityControl/public/pages/aboutView/components/aboutViewHeader.js
